Make SignTool build task functional, remove SignTool.exe (#350)
